bot/requirements.txt
# Nodo - Intra-Platform Arbitrage Bot
# Python 3.10+

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Async support
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Data processing
pandas==2.1.4
numpy==1.26.2

# Environment variables
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Scheduling
schedule==1.2.1
apscheduler==3.10.4

# CLI colors
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Web3 (for execution - install separately if needed)
# pip install web3 eth-account
# Requires Visual C++ Build Tools on Windows
